增加几个基础类

+++ libmx/base/Toolkit.cpp
@@ -5,7 +5,7 @@ BEG_MX_NAMESPACE
 void Toolkit::sleep( long long millisec )
 {
 #ifdef WIN32
-	Sleep(millisec);
+	Sleep((int)millisec);
 #else
 	usleep(millisec * 1000);
 #endif
